+++ devops/setup_networker.sh
@@ -7,5 +7,5 @@ conda config --set always_yes yes --set changeps1 no
 echo "Setup Networker env"
 rm -rf $HOME/miniconda/envs/networker
 conda create -n networker python=2.7
-conda install -n networker -c ioos -c sel networkplanner-metrics
-conda install -n networker -c ioos -c sel networker
+conda install -n networker -c conda-forge -c sel networkplanner-metrics
+conda install -n networker -c conda-forge -c sel networker