+++ modelrunner/redisent/__init__.py
@@ -1,0 +1,11 @@
+"""
+redisent
+
+A package for simple, silent (maybe even 'reticent') management/persistence
+of python entities
+
+If more functionality is needed, see
+[rom](http://pythonhosted.org/rom/)
+"""
+
+from entity import RedisEntity  # noqa