+++ scripts/job_creator.py
@@ -4,27 +4,32 @@
 Script for submitting job to system manually
 """
 
-# setup config options
-from modelrunner import config
-from tornado.options import parse_command_line, parse_config_file
+import modelrunner
+from modelrunner.settings import initialize
+from modelrunner import (
+    config,
+    Job,
+    PrimaryServer
+)
 
-import modelrunner as mr
+# setup config options
+from tornado.options import parse_command_line, parse_config_file
 
 # so we can load config via cmd line args
 parse_command_line()
 parse_config_file(config.options.config_file)
 
+# initialize the global application settings
+initialize(config.options.redis_url)
 
-jm = mr.JobManager(config.options.redis_url,
-                   config.options.primary_url,
-                   config.options.worker_url,
-                   config.options.data_dir,
-                   {})
+primary_server = PrimaryServer(
+                    config.options.primary_url,
+                    config.options.data_dir)
 
-job = mr.Job(config.options.model)
-jm.add_update_job_table(job)
+job = modelrunner.Job(config.options.model)
+Job[job.uuid] = job
 input_fh = open(config.options.input_file, 'r')
-jm.enqueue(job, input_fh.read())
+primary_server.enqueue(job, input_fh.read())
 input_fh.close()
 
 print("created job {}".format(job.uuid))