+++ scripts/start_primary_production.sh
@@ -2,7 +2,6 @@
 
 # start all processes in support of primary server in background
 # assumes modelrunner conda environment has been activated
-redis-server > redis.log 2>&1 & echo $! > redis.pid 
 job_server.py --port=8080 > job_server.log 2>&1 & echo $! > job_server.pid
 job_primary.py > job_primary.log 2>&1 & echo $! > job_primary.pid