+++ testing/test_full.sh
@@ -22,6 +22,14 @@ MR_TEST_SRC_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
 trap mr_cleanup EXIT
 
+# ensure that job list is empty to start
+echo "checking that there are no jobs yet"
+num_jobs=$(mr_get_jobs | python -c "import sys, json; print len(json.load(sys.stdin)[\"data\"])")
+if [ $num_jobs -ne 0 ]
+then
+    exit 1
+fi
+
 # kick off a job
 echo "creating new job"
 job_id=$(mr_create_job test_full_`date +%Y-%m-%d_%H:%M:%S` "test" "@testing/sleep_count_8.zip")
@@ -41,9 +49,9 @@ mr_wait_for_status $job_kill_id "RUNNING" 2
 echo "attempting to kill job $job_kill_id"
 mr_kill_job $job_kill_id
 
-# ensure that eventually the status of the killed job goes to "FAILED"
-echo "waiting for job $job_kill_id to go FAILED"
-mr_wait_for_status $job_kill_id "FAILED" 2
+# ensure that eventually the status of the killed job goes to "KILLED"
+echo "waiting for job $job_kill_id to go KILLED"
+mr_wait_for_status $job_kill_id "KILLED" 2
 
 # ensure that we can kill 2nd job while 1st job is running
 echo "creating long running job"
@@ -59,18 +67,23 @@ mr_wait_for_status $job_2_id "RUNNING" 2
 echo "attempting to kill job $job_2_id"
 mr_kill_job $job_2_id
 
-# ensure that eventually the status of the killed job goes to "FAILED"
-echo "waiting for job $job_2_id to go FAILED"
-mr_wait_for_status $job_2_id "FAILED" 2
+echo "waiting for job $job_2_id to go KILLED"
+mr_wait_for_status $job_2_id "KILLED" 2
 
 # test killing long running job
 echo "attempting to kill job $job_long_id"
 mr_kill_job $job_long_id
 
-# ensure that eventually the status of the killed job goes to "FAILED"
-echo "waiting for job $job_long_id to go FAILED"
-mr_wait_for_status $job_long_id "FAILED" 2
+echo "waiting for job $job_long_id to go KILLED"
+mr_wait_for_status $job_long_id "KILLED" 2
 
+# we should have 4 jobs now
+echo "checking that there are 4 jobs"
+num_jobs=$(mr_get_jobs | python -c "import sys, json; print len(json.load(sys.stdin)[\"data\"])")
+if [ $num_jobs -ne 4 ]
+then
+    exit 1
+fi
 
 # disable the trap now
 trap - EXIT